Revert "Send stability in calibration stage & FridaInstrumentationHelper retunrs Result<Self, Error> (#1056)"
This reverts commit 4d78878c02846b2c0a49686cd05cccadd2b0ac72.
