+++ libafl/src/events/mod.rs
@@ -3,7 +3,11 @@
 pub mod simple;
 pub use simple::*;
 pub mod llmp;
-use alloc::{boxed::Box, string::String, vec::Vec};
+use alloc::{
+    boxed::Box,
+    string::{String, ToString},
+    vec::Vec,
+};
 #[cfg(all(unix, feature = "std"))]
 use core::ffi::c_void;
 use core::{
@@ -29,6 +33,7 @@ use crate::{
     inputs::Input,
     monitors::UserStats,
     observers::ObserversTuple,
+    stages::calibrate::UnstableEntriesMetadata,
     state::{HasClientPerfMonitor, HasExecutions, HasMetadata},
     Error,
 };
@@ -456,6 +461,23 @@ where
                     phantom: PhantomData,
                 },
             )?;
+
+            // Send the stability event to the broker
+            if let Some(meta) = state.metadata().get::<UnstableEntriesMetadata>() {
+                let unstable_entries = meta.unstable_entries().len();
+                let map_len = meta.map_len();
+                self.fire(
+                    state,
+                    Event::UpdateUserStats {
+                        name: "stability".to_string(),
+                        value: UserStats::Ratio(
+                            (map_len - unstable_entries) as u64,
+                            map_len as u64,
+                        ),
+                        phantom: PhantomData,
+                    },
+                )?;
+            }
 
             // If performance monitor are requested, fire the `UpdatePerfMonitor` event
             #[cfg(feature = "introspection")]
