+++ libafl/src/stages/calibrate.rs
@@ -13,7 +13,7 @@ use serde::{Deserialize, Serialize};
 use crate::{
     bolts::{current_time, tuples::Named, AsIter},
     corpus::{Corpus, CorpusId, SchedulerTestcaseMetaData},
-    events::{Event, EventFirer, LogSeverity},
+    events::{EventFirer, LogSeverity},
     executors::{Executor, ExitKind, HasObservers},
     feedbacks::{
         map::{IsNovel, MapFeedback, MapFeedbackMetadata, Reducer},
@@ -21,7 +21,6 @@ use crate::{
     },
     fuzzer::Evaluator,
     inputs::UsesInput,
-    monitors::UserStats,
     observers::{MapObserver, ObserversTuple},
     schedulers::powersched::SchedulerMetadata,
     stages::Stage,
@@ -284,20 +283,6 @@ where
 
             data.set_bitmap_size(bitmap_size);
             data.set_handicap(handicap);
-        }
-
-        // Send the stability event to the broker
-        if let Some(meta) = state.metadata().get::<UnstableEntriesMetadata>() {
-            let unstable_entries = meta.unstable_entries().len();
-            let map_len = meta.map_len();
-            mgr.fire(
-                state,
-                Event::UpdateUserStats {
-                    name: "stability".to_string(),
-                    value: UserStats::Ratio((map_len - unstable_entries) as u64, map_len as u64),
-                    phantom: PhantomData,
-                },
-            )?;
         }
 
         Ok(())
